(feature): Added more guard clauses and removal options to uninstall script

+++ uninstall.sh
@@ -21,16 +21,41 @@ uninstall_python
 rm -fR .
 
 docker rmi nginx
-docker rmi postgres:15
 docker rmi docker-test-green-coding-gunicorn
 docker rmi docker-green-coding-gunicorn
+
+read -p "Do you also want to remove the database and your data? (y/N) : " remove_db
+if [[  "$remove_db" == "Y" || "$remove_db" == "y" ]] ; then
+    docker rmi postgres:15
+    docker volume rm docker_green-coding-postgres-data
+fi
+
 docker system prune
+
 
 if [[ $(uname) == "Linux" ]]; then
     if lsb_release -is | grep -q "Fedora"; then
-        sudo dnf -y install msr-tools lm_sensors lm_sensors-devel glib2 glib2-devel tinyproxy stress-ng lshw ipmitool
+        sudo dnf -y remove msr-tools lm_sensors lm_sensors-devel glib2 glib2-devel tinyproxy stress-ng lshw ipmitool
     else
-        sudo apt remove lm-sensors libsensors-dev libglib2.0-0 libglib2.0-dev tinyproxy stress-ng lshw freeipmi-tools ipmitool msr-tools -y
+        sudo apt remove -y lm-sensors libsensors-dev libglib2.0-0 libglib2.0-dev tinyproxy stress-ng lshw freeipmi-tools ipmitool msr-tools
+    fi
+
+    read -p "Do you also want to remove pre-install requirements (curl git make gcc python3 python3-devel)? (y/N) : " pre_install
+    if [[  "$pre_install" == "Y" || "$pre_install" == "y" ]] ; then
+        if lsb_release -is | grep -q "Fedora"; then
+            sudo dnf remove -y curl git make gcc python3 python3-devel
+        else
+            sudo apt remove -y git make gcc python3 python3-pip python3-venv
+        fi
+    fi
+
+    read -p "Do you also want to remove pre-install requirements (docker)? (y/N) : " pre_docker
+    if [[  "$pre_docker" == "Y" || "$pre_docker" == "y" ]] ; then
+        if lsb_release -is | grep -q "Fedora"; then
+            sudo dnf remove -y docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
+        else
+            sudo apt remove -y docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
+        fi
     fi
 fi
